# Avoid passing variable size to memcpy to prevent external call