@flobernd 2022-05-05
In the decoder we have a similar function that read bytes. It was well uses 'memcpy' so maybe we can optimize it was well.