@ZehMatt 2022-05-05
> In the decoder we have a similar function that read bytes. It was well uses 'memcpy' so maybe we can optimize it was well.

You probably mean ZydisInputNextBytes, I need to look into that. This one is quite easy to reason with as the size constraint is limited to 4 values, I'll have to see how ZydisInputNextBytes is used in general but we can probably go with common integer sizes at least until we fallback to memcpy. Will check it out.