@flobernd 2022-05-06
Yes, that's the function I had in mind. If I remember correctly, it can read 2, 4 and 8 bytes (but I will double check).

For now, we can merge this PR if @athre0z is fine with that too.