@flobernd 2022-05-06
@athre0z We can as well wrap it in a `#ifdef ZYAN_WINDOWS` to mitigate the regression on Linux.