@athre0z 2022-05-06
Not worth the complexity IMO.